requirements/base.txt
bcrypt==3.1.3
boto3==1.4.4
cryptography==1.8.1
https://github.com/3Blades/django/archive/1.11-uuid-pk.zip
django-filter==1.0.4
django-rest-swagger==2.1.2
djangorestframework==3.6.2
docker==2.5.1
drf-nested-routers==0.90.0
GitPython==2.1.1
hashids==1.2.0
hiredis==0.2.0
humanize==0.5.1
nameparser==0.5.1
passlib==1.7.0
psycopg2==2.7.1
Pygments==2.2.0
python-dateutil==2.6.0
python-magic==0.4.13
pytz==2017.2
rauth==0.7.3
dj-database-url==0.4.2
drf-ujson==1.2
ujson==1.35
django-redis==4.8.0
django-extensions==1.7.9
gunicorn==19.7.1
celery==4.1.0
django-celery-results==1.0.1
uuid==1.30
django-cacheops==3.2.1
django-storages==1.5.2
django-oauth-toolkit==0.12.0
django-rest-framework-social-oauth2==1.0.6
djangorestframework-jwt==1.10
slackclient==1.0.5
django-cors-headers==2.0.2
django-guardian==1.4.8
Pillow==4.2.1
stripe==1.55.2
django-haystack-elasticsearch==0.1.0
elasticsearch>=5.0.0,<6.0.0
drf-haystack==1.6.1
Pillow==4.2.1
djoser==0.6.0
boto==2.48.0
django-ses==0.8.2
google-compute-engine==2.4.1
raven==6.2.1
celery-redbeat==0.10.0
django-treebeard==4.1.2
